bugfix / work-around for autocomplete bug
This fixes a bug in the UI, when selecting results from an autocomplete
field in a relation form, the selected value is not correctly parsed.

This is due to intermixing human and machine data, because the human
presentation string was used as identifier. I fixed it by squeezing
human and machine data into one string, which is separated later on
again. It's a hack and I'm not proud of it, but I didn't want to touch
too much legacy code and risk further regressions.

This approach seems economic and least invasive. Especially so since
this is a legacy codebase anyway.